UI Redesign/Feature Add
Changed ListView to DataGridView. This allows for copy-pasting as well as editing of cells. 

Broke the right click menu of the DataGrid, but working on fixing that. 

Also introduced cursor Point and Time to Sleep Variance to allow for more "random", human-like interaction. After all, a human is a lot less likely to click at exactly every 1000 ms on the same pixel/set of pixels...